various more cleanup
