+++ src/runtime/worker_handle.rs
@@ -31,10 +31,10 @@ impl WorkerHandleListener {
 
     /// Returns whether there are handles that have not been dropped yet.
     pub(crate) fn check_if_alive(&mut self) -> bool {
-        !matches!(
-            self.receiver.try_recv(),
-            Err(mpsc::error::TryRecvError::Disconnected)
-        )
+        match self.receiver.try_recv() {
+            Err(mpsc::error::TryRecvError::Disconnected) => false,
+            _ => true,
+        }
     }
 
     /// Constructs a new channel for for monitoring whether this worker still has references
